+++ UKParliament/Models/LocatedSignatureCount.cs
@@ -15,7 +15,7 @@ namespace UKParliament.Model
 
         public Place LocatedSignatureCountHasPlace => ((IEnumerable<Place>)new DynamicObjectCollection<Place>(this, Property(UKParliamentOntology.LocatedSignatureCountHasPlace.Uri))).SingleOrDefault();
 
-        public ICollection<int> SignatureCount => new DynamicObjectCollection<int>(this, Property(UKParliamentOntology.SignatureCount.Uri));
+        public ICollection<System.Int64> SignatureCount => new DynamicObjectCollection<System.Int64>(this, Property(UKParliamentOntology.SignatureCount.Uri));
 
         public ICollection<string> SignatureCountRetrievedAt => new DynamicObjectCollection<string>(this, Property(UKParliamentOntology.SignatureCountRetrievedAt.Uri));
     }
